supporting multiple test runs
use the --multiple or -m arugment to save each test suite as a separate xml file.

the default mode is to overwrite any existing test suite xml file as the
suite is completed. the net effect is that only the results
from the final destination are reflected in the output files.

this usually occurs when using xcodebuild with multiple -destination
arguments...eg. running specs on several simulators or devices.